wbjt/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
cramjam==2.7.0
fastparquet==0.8.1
fsspec==2023.1.0
idna==3.4
numpy==1.26.0
pandas==2.1.1
py4j==0.10.9.7
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tzdata==2023.3
urllib3==1.26.6
